+++ src/main/java/com/example/myShop/domain/dto/goods/GoodsUpdateDto.java
@@ -1,14 +1,11 @@
 package com.example.myShop.domain.dto.goods;
 
-import com.example.myShop.domain.dto.category.CategoryDto;
-import com.example.myShop.domain.dto.producer.ProducerDto;
-import com.example.myShop.domain.entity.Order;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Value;
 import lombok.extern.jackson.Jacksonized;
 
-import java.util.List;
+import java.math.BigDecimal;
 
 import static lombok.AccessLevel.PRIVATE;
 
@@ -23,7 +20,7 @@ import static lombok.AccessLevel.PRIVATE;
 @AllArgsConstructor(access = PRIVATE)
 public class GoodsUpdateDto {
         String name;
-        int price;
-        int availability;
+        BigDecimal price;
+        long availability;
         String imageUrl;
 }
